Rewrite initial bookmark resolution algorithm
This combines episode, metadata and archive extraction into a single worker task, so that they may each use the data found in each extraction method for cases where one method is superior to the other.

For example, sometimes the episode extraction can get data hidden from the HTML document extractor. In these cases, we should use that data over the HTML extracted data.

Also rewrites a ton of types and refactors a bunch of things.